Ts-/Webpack config; Remove babel config;
