+++ .npmignore
@@ -4,7 +4,7 @@ stories/
 coverage
 lib/examples/
 .gitignore
-prettier.config.js
+prettier.config.ts
 package-lock.json
 tsconfig.json
 tsconfig.test.json
