+++ .storybook/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../tsconfig",
+  "compilerOptions": {
+    "jsx": "react",
+    "isolatedModules": false,
+    "noEmit": false
+  }
+}
