+++ package.json
@@ -8,6 +8,7 @@
   "types": "lib/index",
   "scripts": {
     "start": "start-storybook -p 6006",
+    "start-strong": "NODE_OPTIONS=\"--max-old-space-size=8192\" start-storybook -p 6006",
     "start-from-scratch": "rm -rf node_modules && yarn && yarn build-index && yarn build-lib && yarn start",
     "prepublishOnly": "yarn && yarn build-index && yarn lint && CI=true yarn test && yarn build-lib",
     "deploy-storybook": "storybook-to-ghpages",
@@ -67,7 +68,7 @@
     "react-docgen-typescript-loader": "^2.2.0",
     "react-docgen-typescript-webpack-plugin": "^1.1.0",
     "react-dom": "16.8.3",
-    "react-scripts": "^2.1.2",
+    "react-scripts": "^2.1.8",
     "react-select": "2.1.1",
     "recompose": "0.30.0",
     "ts-jest": "^23.10.5",
@@ -100,10 +101,7 @@
     "@types/storybook__addon-notes": "^4.0.2",
     "@types/styled-system": "^3.2.2",
     "date-fns": "2.0.0-alpha.26",
-    "lodash": "4.17.10",
-    "react-click-outside": "^3.0.1",
-    "react-popper": "^1.0.2",
-    "react-spinkit": "^3.0.0",
+    "fork-ts-checker-webpack-plugin-alt": "^0.4.14",
     "lodash": "4.17.10",
     "react": "16.8.3",
     "react-click-outside": "^3.0.1",
