+++ src/components/ui/box/ArrowBox.tsx
@@ -11,9 +11,6 @@ export interface ArrowBoxProps {
 export class ArrowBox extends React.Component<ArrowBoxProps> {
   render() {
     const { direction = 'up' } = this.props;
-    console.log('________', {
-      type: Object as () => ArrowBoxProps
-    });
     return (
       <ArrowWrapper>
         <div className={`arrow_box ${direction}`}>{this.props.children}</div>
