+++ tsconfig.json
@@ -34,7 +34,7 @@
     "skipLibCheck": true,
     "esModuleInterop": false,
     "allowSyntheticDefaultImports": true,
-    "isolatedModules": true,
+    "isolatedModules": false,
     "noEmit": true
   },
   "include": [
